DDCNL-3814: Created Filtered case class seperate from Exclusions
