+++ app/uk/gov/hmrc/nisp/controllers/StatePensionController.scala
@@ -168,7 +168,7 @@ class StatePensionController @Inject()(authenticate: AuthAction,
               auditConnector.sendEvent(AccountExclusionEvent(
                 user.nino.nino,
                 user.name,
-                statePensionExclusion
+                statePensionExclusion.exclusion
               ))
               Redirect(routes.ExclusionController.showSP).withSession(storeUserInfoInSession(user, contractedOut = false))
             case _ => throw new RuntimeException("StatePensionController: SP and NIR are unmatchable. This is probably a logic error.")
